added lifebar to new enemy prefabs
