+++ Assets/Scripts/enemies/HealthEnemy.cs
@@ -1,15 +1,39 @@
 using UnityEngine;
+using UnityEngine.UI;
 
 /// <summary>
 ///     Manages the health of enemies including taking damage and deleting killed ennemies
 /// </summary>
 public class EnemyHealth : MonoBehaviour
 {
-    [Header("Attributes")]
+    #region attributes
     [SerializeField] private int hitPoints = 2;
     [SerializeField] private int currencyWorth;
+    [SerializeField] private Image lifebarFill;
 
+    private Camera cam;
+    private int maxHitPoints;
     private bool isDestroyed = false;
+    #endregion
+
+    void Start()
+    {
+        maxHitPoints = hitPoints;
+
+        // set main camera as world space camera
+        cam = Camera.main;
+        Canvas lifeBar = transform.GetChild(0).GetComponent<Canvas>();
+
+        if (lifeBar != null && lifeBar.renderMode == RenderMode.WorldSpace)
+        {
+            lifeBar.worldCamera = Camera.main;
+        }
+    }
+
+    void Update()
+    {
+        UpdateLifeBar();
+    }
 
     /// <summary>
     /// Deals damage to the enemy by decreasing his hit points.
@@ -29,8 +53,13 @@ public class EnemyHealth : MonoBehaviour
         }
     }
 
-    public int GetHitpoints()
+    /// <summary>
+    ///     Updates the lifebar above enemies by displaying the remaining hit points
+    /// </summary>
+    private void UpdateLifeBar()
     {
-        return hitPoints;
+        lifebarFill.fillAmount = (float)hitPoints / maxHitPoints;
+        transform.rotation = cam.transform.rotation;
     }
+
 }
