--- Assets/Scripts/enemies/Lifebar.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: de11a3e2e78e8374893f631a71424691
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
